spring-core/src/main/java/com/techspring/spring/core/config/PrimaryConfiguration.java
package com.techspring.spring.core.config;

import com.techspring.spring.core.model.Foo;
import org.springframework.context.annotation.Bean;
import org.springframework.context.annotation.Configuration;
import org.springframework.context.annotation.Primary;

@Configuration
public class PrimaryConfiguration {

    @Primary //nilai primary menjadi utama, otomatis akan mengambil nilai @Primary
    @Bean
    public Foo foo1(){
        return new Foo();
    }

    @Bean
    public Foo foo2(){
        return new Foo();
    }
}
